+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 2025-08-01
+
+- Added `.run [code]` command to execute JavaScript code directly without creating an alias.
+
 ## 2025-07-31
 
 - Added docs, see [docs/README.md](docs/README.md).